#96 refactor: changed search to search position score instead of search action score (simplified code)
